Fully implement add, delete endpoints

+++ DocAssistantWebApi/Controllers/PatientController.cs
@@ -1,8 +1,11 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using DocAssistant_Common.Models;
 using DocAssistantWebApi.Database.Repositories;
+using DocAssistantWebApi.Errors;
 using DocAssistantWebApi.Filters;
 using DocAssistantWebApi.Services.Auth;
 using Microsoft.AspNetCore.Authorization;
@@ -15,11 +18,13 @@ namespace DocAssistantWebApi.Controllers
 
         private readonly IRepository<Patient> _patientRepository;
         private readonly IRepository<Doctor> _doctorRepository;
+        private readonly IRepository<Assistant> _assistantRepository;
 
-        public PatientController(IRepository<Patient> repository,IRepository<Doctor> doctorRepository)
+        public PatientController(IRepository<Patient> repository,IRepository<Doctor> doctorRepository,IRepository<Assistant> assistantRepository)
         {
             this._patientRepository = repository;
             this._doctorRepository = doctorRepository;
+            this._assistantRepository = assistantRepository;
         }
 
         [Authorize(Policy = "DoctorRequirement")]
@@ -28,7 +33,6 @@ namespace DocAssistantWebApi.Controllers
         [HttpPost]
         public async Task<ActionResult> LoadPatient([FromQuery(Name = "type")] string type, [FromQuery(Name = "id")] long? id)
         {
-
             long docId = (long) HttpContext.Items["Id"];
             
             switch (type)
@@ -38,8 +42,13 @@ namespace DocAssistantWebApi.Controllers
                     var patient = await this._patientRepository.Where(patient => patient.Id == id);
                     if (patient == null || patient.DoctorId != docId)
                     {
-                        return BadRequest(
-                            "Patient with id does not exist or the patient does not belong to this doctor account");
+                        throw new GenericRequestException
+                        {
+                            Title = "Failed to load patient",
+                            Error =
+                                "Patient with id does not exist or the patient does not belong to this doctor account",
+                            StatusCode = 400
+                        };
                     }
                     return Ok(patient);
                 }
@@ -48,7 +57,12 @@ namespace DocAssistantWebApi.Controllers
                         await this._patientRepository.WhereMulti(patient => patient.DoctorId == docId);
                     return Ok(patients);
                 default:
-                    return BadRequest();
+                    throw new GenericRequestException
+                    {
+                        Title = "Failed to load patient",
+                        Error = "Invalid request type",
+                        StatusCode = 400
+                    };
             }
         }
 
@@ -56,13 +70,45 @@ namespace DocAssistantWebApi.Controllers
         [Produces("application/json")]
         [Route("api/patient")]
         [HttpPost]
-        public async Task<ActionResult> AddPatient([FromHeader(Name = "Authorization")] string token,[FromBody] Patient patient)
+        public async Task<ActionResult> AddPatient([FromBody] Patient patient)
         {
-            long docId = (long)HttpContext.Items["Id"];
+            var accountType = (Roles)HttpContext.Items["AccountType"];
             
+            long docId;
+
+            if (accountType == Roles.Assistant)
+            {
+                var assistant =
+                    await this._assistantRepository.Where(assistant => assistant.Id == (long) HttpContext.Items["Id"]);
+
+                docId = assistant.DoctorId;
+            }
+            else
+                docId = (long)HttpContext.Items["Id"];
+
+            var isSSNUnique = await this._patientRepository.Where(entity => entity.SSN == patient.SSN) == null;
+            if (!isSSNUnique)
+            {
+                throw new GenericRequestException
+                {
+                    Title = "Failed to add patient",
+                    Error = "SSN must be unique",
+                    StatusCode = 400
+                };
+            }
+            
+            patient.ArriveTime = DateTime.Now;
+
             var doctor = await this._doctorRepository.Where(doctor => doctor.Id == docId);
             doctor.Patients.Add(patient);
-            await this._doctorRepository.Update(doctor);
+            
+            if(!await this._doctorRepository.Update(doctor))
+                throw new GenericRequestException
+                {
+                    Title = "Failed to add patient",
+                    Error = "Save failed",
+                    StatusCode = 400
+                };
 
             return Ok();
         }
@@ -73,8 +119,31 @@ namespace DocAssistantWebApi.Controllers
         [HttpPatch]
         public async Task<ActionResult> UpdateData([FromBody] Patient patient)
         {
-            await this._patientRepository.Update(patient);
+            var check = await this._patientRepository.Where(entity => entity.Id == patient.Id);
+            if(check == null)
+                throw new GenericRequestException
+                {
+                    Title = "Failed to update patient data",
+                    Error = "Patient does not exist",
+                    StatusCode = 400
+                };
             
+            if(check.DoctorId != (long)HttpContext.Items["Id"])
+                throw new GenericRequestException
+                {
+                    Title = "Failed to update patient data",
+                    Error = "Patient belongs to a different doctor",
+                    StatusCode = 400
+                };
+
+            if(!await this._patientRepository.Update(patient)) 
+                throw new GenericRequestException
+                {
+                    Title = "Failed to update patient data",
+                    Error = "No changes were made",
+                    StatusCode = 400
+                };
+
             return Ok();
         }
 
@@ -84,7 +153,34 @@ namespace DocAssistantWebApi.Controllers
         [HttpDelete]
         public async Task<ActionResult> DeletePatient([FromQuery(Name = "id")] long id)
         {
-            await this._patientRepository.DeleteWhere(patient => patient.Id == id);
+            var check = await this._patientRepository.Where(entity => entity.Id == id);
+            if(check == null)
+                throw new GenericRequestException
+                {
+                    Title = "Failed to delete patient data",
+                    Error = "Patient does not exist",
+                    StatusCode = 400
+                };
+            
+            if(check.DoctorId != (long)HttpContext.Items["Id"])
+                throw new GenericRequestException
+                {
+                    Title = "Failed to delete patient data",
+                    Error = "Patient belongs to a different doctor",
+                    StatusCode = 400
+                };
+            
+            int count = await this._patientRepository.DeleteWhere(patient => patient.Id == id);
+
+            if (count == 0)
+            {
+                throw new GenericRequestException
+                {
+                    Title = "Failed to delete patient data",
+                    Error = "Patient does not exist",
+                    StatusCode = 400
+                };
+            }
 
             return Ok();
         }
